backend/requirements.txt
Flask
flask-cors
flask-restful
torch
torchvision
Pillow
numpy
